pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "concord-sc"
version = "0.1.0"
description = "Concord single-cell foundation model with collaborative rotary attention."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pyyaml>=6.0",
  "torch>=2.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
singlecell = [
  "anndata>=0.10",
  "scipy>=1.10",
]
flash = [
  "flash-attn>=2.5",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
